backend/pyproject.toml
[project]
name = "assetlog-backend"
version = "0.1.0"
description = "AssetLog FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncmy>=0.2.9",
    "alembic>=1.13.3",
    "pydantic[email]>=2.9.2",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "httpx>=0.27.2",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.1",
    # Price adapter libraries — included now, imported in Step 2+
    "yfinance>=0.2.48",
    "pykrx>=1.0.51",
    "finance-datareader>=0.9.96",
    "ccxt>=4.4.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.2",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-python-jose>=3.3.4",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=xml:coverage.xml"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long — handled by formatter
    "B008",  # do not perform function calls in default args (FastAPI Depends pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
# pytest fixtures use untyped patterns that conflict with strict mode
disallow_untyped_defs = false
disallow_untyped_decorators = false
warn_unused_ignores = false

[tool.coverage.run]
source = ["app"]
omit = ["app/db/base.py"]

[tool.coverage.report]
show_missing = true
fail_under = 0
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "mypy>=1.20.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
    "types-python-jose>=3.5.0.20260408",
]
